ISSUE #1976 - Make filter list scrollable
